Major refactor, implemented proper folder structure and deleted redundant files
