--- src/AdminAuth.css
@@ -1,4 +1,0 @@
-.floating-icon {
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    text-align: center;
-  }
