--- src/NavBar.css
@@ -1,3 +1,0 @@
-.whiteStyle {
-    color: white;
-}
